+++ src/lint/eslint.js
@@ -1,6 +1,14 @@
-/**
- * Shareable ESLint config -- stub.
- */
 'use strict';
 
-module.exports = {};
+const wordpress = require('@wordpress/eslint-plugin');
+const commentsConfigs = require('@eslint-community/eslint-plugin-eslint-comments/configs');
+const jestPlugin = require('eslint-plugin-jest');
+
+module.exports = [
+	...wordpress.configs.recommended,
+	commentsConfigs.recommended,
+	{
+		...jestPlugin.configs['flat/recommended'],
+		files: ['**/*.test.js'],
+	},
+];